Refactored generic plot generation (#1524)
* added defined interfaces for adding new quantities
* added documentation both general, and function specific

Fix #1522
